1 Исправил методы сохранения и обносления записей в БД 3 дописал связи Organization и service 2 написал тесты для DAO

+++ src/main/java/ru/bellintegrator/practice/countries/dao/CountriesDao.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface CountriesDao {
     List<Countries> all();
-    Countries loadById (Long id);
+    Countries loadById (Integer id);
     void save(Countries countries);
 }
