+++ src/main/java/ru/bellintegrator/practice/user/dao/UserDao.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface UserDao {
     List<User> all();
     User loadByID(Long id);
-    List<User> GetFilter(Long officeid, String firstname,String second_name, String middle_name,String position, Integer doc_code , Integer citizenship_code  );
+    List<User> GetFilter(Long officeid, String firstname,String second_name, String middle_name,String position, Long doc_code , Integer citizenship_code  );
 
     void save (User user);
     void update(UserView user);
